+++ tracing/jaeger/tracer.go
@@ -18,6 +18,10 @@ func NewTracer(params TracerParams) (opentracing.Tracer, error) {
 		jaegerOptions = append(jaegerOptions, jaegercfg.Logger(&kitLogger{params.Logger}))
 	}
 
+	if params.MetricsFactory != nil {
+		jaegerOptions = append(jaegerOptions, jaegercfg.Metrics(params.MetricsFactory))
+	}
+
 	// TODO: handle closer
 	tracer, closer, err := params.Config.JaegerConfig.New(params.Config.ServiceName, jaegerOptions...)
 	if err != nil {