+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Upgrade Jaeger libraries
 - Use go-kit logger wrapper from jaeger lib
+- Upgraded gRPC
 
 
 ## [0.7.0] - 2018-01-12