+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+
+## [0.8.0] - 2018-01-14
+
 ### Added
 
 - Metrics support for Jaeger
@@ -185,7 +188,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Bootstrap options
 
 
-[Unreleased]: https://github.com/goph/fxt/compare/v0.7.0...HEAD
+[Unreleased]: https://github.com/goph/fxt/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/goph/fxt/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/goph/fxt/compare/v0.6.3...v0.7.0
 [0.6.3]: https://github.com/goph/fxt/compare/v0.6.2...v0.6.3
 [0.6.2]: https://github.com/goph/fxt/compare/v0.6.1...v0.6.2